change test file name
